move paths utils to private module + indentation in doc
